commands: move commit code to commit.rs
